rename interface for stroage and add implentation in another package
